+++ src/components/DownloadAndShareComponents/helpers/getDownloadResourceType.tsx
@@ -3,18 +3,18 @@ import {
   PreselectedDownloadType,
   PreselectedShareType,
   ShareResourceType,
-  preselectedDownloadTypeMap,
-  preselectedShareTypeMap,
-} from "../downloadsAndShare.types";
+} from "../downloadAndShare.types";
+
+import { combinedPreselectedTypeMap } from "./combinedPreselectedTypeMap";
 
 export const getPreselectedDownloadResourceTypes = (
   title: PreselectedDownloadType,
 ): DownloadResourceType[] | "all" | undefined => {
-  return preselectedDownloadTypeMap[title];
+  return combinedPreselectedTypeMap[title].downloadType;
 };
 
 export const getPreselectedShareResourceTypes = (
   title: PreselectedShareType,
 ): ShareResourceType[] | "all" | undefined => {
-  return preselectedShareTypeMap[title];
+  return combinedPreselectedTypeMap[title].shareType;
 };